Fix for serialized focal pointes with new tests.

+++ thumbor/point.py
@@ -27,7 +27,7 @@ class FocalPoint(object):
 
     @classmethod
     def from_dict(cls, values):
-        return cls(int(values['x']), int(values['y']), int(values['z']))
+        return cls(float(values['x']), float(values['y']), float(values['z']))
 
     def __init__(self, x, y, weight=1.0):
         self.x = x
